+++ ark/schema/shared/intersections.ts
@@ -5,7 +5,6 @@ import type { Morph } from "../roots/morph.ts"
 import type { BaseRoot } from "../roots/root.ts"
 import type { Union } from "../roots/union.ts"
 import type { BaseScope } from "../scope.ts"
-import type { NodeMeta } from "./declare.ts"
 import { Disjoint } from "./disjoint.ts"
 import {
 	rootKinds,
@@ -149,7 +148,7 @@ const pipeMorphed = (
 
 			// otherwise, the input has not changed so preserve metadata
 
-			let meta: NodeMeta | undefined
+			let meta: ArkEnv.meta | undefined
 
 			if (viableBranches.length === 1) {
 				const onlyBranch = viableBranches[0]