+++ ark/type/parser/shift/operand/operand.ts
@@ -13,15 +13,19 @@ import { parseUnenclosed, writeMissingOperandMessage } from "./unenclosed.ts"
 
 export const parseOperand = (s: RuntimeState): void =>
 	s.scanner.lookahead === "" ? s.error(writeMissingOperandMessage(s))
-	: s.scanner.lookahead === "(" ? s.shiftedByOne().reduceGroupOpen()
+	: s.scanner.lookahead === "(" ? s.shiftedBy(1).reduceGroupOpen()
 	: s.scanner.lookaheadIsIn(enclosingChar) ? parseEnclosed(s, s.scanner.shift())
-	: s.scanner.lookaheadIsIn(whitespaceChars) ? parseOperand(s.shiftedByOne())
+	: s.scanner.lookaheadIsIn(whitespaceChars) ? parseOperand(s.shiftedBy(1))
 	: s.scanner.lookahead === "d" ?
 		s.scanner.nextLookahead in enclosingQuote ?
 			parseEnclosed(
 				s,
 				`${s.scanner.shift()}${s.scanner.shift()}` as EnclosingStartToken
 			)
+		:	parseUnenclosed(s)
+	: s.scanner.lookahead === "x" ?
+		s.scanner.nextLookahead === "/" ?
+			s.shiftedBy(2) && parseEnclosed(s, "x/")
 		:	parseUnenclosed(s)
 	:	parseUnenclosed(s)
 
@@ -40,6 +44,10 @@ export type parseOperand<s extends StaticState, $, args> =
 				>
 			) ?
 				parseEnclosed<s, `d${enclosing}`, nextUnscanned>
+			:	parseUnenclosed<s, $, args>
+		: lookahead extends "x" ?
+			unscanned extends Scanner.shift<"/", infer nextUnscanned> ?
+				parseEnclosed<s, "x/", nextUnscanned>
 			:	parseUnenclosed<s, $, args>
 		:	parseUnenclosed<s, $, args>
 	:	s.completion<`${s["scanned"]}${BaseCompletions<$, args>}`>